Added basic support for Linux.
Currently only sends dummy text to the previously active window.
Uses xdotool to get and minimize the active window, thus leaving the previous
as the active window.

Build and include files updated to include proper files based on OS.

+++ csvdataparser.hpp
@@ -34,7 +34,7 @@ public:
             fb.close();
 
             mHeaders = data[0];
-            for (int n=1; n<data.size(); n++)
+            for (size_t n=1; n<data.size(); n++)
             {
                 mRowDataMap[n-1] = data[n];
             }
